+++ inst/tinytest/test_helpers.R
@@ -1,0 +1,6 @@
+# Odd integer ==================================================================
+x <- 2
+expect_error(alkahest:::assert_odd(x))
+
+x <- 3
+expect_equal(alkahest:::assert_odd(x), x)