+++ NEWS.md
@@ -6,6 +6,9 @@
 ## Bugfixes & changes
 * `baseline_peakfilling()` now smoothes the data using the Whittaker smoother as defined in Liland (2015).
 
+## Internals
+* Use **tinytest** instead of **testthat**.
+
 # alkahest 1.0.0
 
 * First release